added facial keypoints as well as segmentation

+++ exercise_4/exercise_code/data_utils.py
@@ -1,11 +1,13 @@
 """Data utility functions."""
 import os
 
+import matplotlib.pyplot as plt
 import numpy as np
 import torch
 import torch.utils.data as data
 from PIL import Image
 from torchvision import transforms
+
 import _pickle as pickle
 
 # pylint: disable=C0326
@@ -237,6 +239,7 @@ def rel_error(x, y):
     assert x.shape == y.shape, "tensors do not have the same shape. %s != %s" % (x.shape, y.shape)
     return np.max(np.abs(x - y) / (np.maximum(1e-8, np.abs(x) + np.abs(y))))
 
+
 def string2image(string):
     """Converts a string to a numpy array."""
     return np.array([int(item) for item in string.split()]).reshape((96, 96))
@@ -247,5 +250,6 @@ def get_image(idx, key_pts_frame):
     
 def get_keypoints(idx, key_pts_frame):
     keypoint_cols = list(key_pts_frame.columns)[:-1]
-    return key_pts_frame.iloc[idx][keypoint_cols].values.reshape((15, 2))
+    return key_pts_frame.iloc[idx][keypoint_cols].values.reshape((15, 2)).astype(np.float32)
   
+
